# Speed up general character class matching